Disable tmux integration in powerline.zsh bindings
The integration was annoingly setting TMUX_PWD_N and TMUX_COLUMNS_N
variables every time a new tmux window/pane was opened, without cleaning
up on thier closure, ending up filling the environment with hundreds
of variables if you used the session for long.

Since the integration is not used to display tmux info in the zsh
prompt, it should be safe to disable.