Remove workaround in zshenv
This was required because this of this option in tmux:
    set-option -g default-command "reattach-to-user-namespace -l zsh" causes

The option was causing .zshenv to be loaded twice, which is not the case
anymore. The option itself was required to make vim clipboard register
work properly on Mac OS X, but that seems to work fine without it now.